+++ src/Using.cs
@@ -1,3 +1,4 @@
 global using Microsoft.JSInterop;
 global using Blazor.Avatar.Interop;
 global using Blazor.Avatar.Components.InitialAvatarComponent;
+global using Blazor.Avatar.Components.IdenticonAvatarComponent;
